Remove call to controlClient.GetServiceEndpoints
Fixes ZEN-13511
facade.service.GetServiceEndpoints was throwing an error, due to the fact
that it was filtering on service states within the same pool as this
service, and this service is running in a separate pool.  Because
agent_proxy.GetServiceEndpoints was exiting early (due to the error), we
were never getting a list of the control-plane specific endpoints in
handleControlCenterImports, which is all it needs.  The non-control-plane
specific results were being filtered out anyway, so there is no need to
gather them.