CC-1317: serviced wait for network to come up
In our systemctl unit definition we wait for network-online.target, but
that condition appears to be satisfied when any non-loopback interface
gets an IP address?  What we constantly see in our log is docker0 has an
IP address but the NIC card does not, and our code panics.

Instead of panicing here, we'll loop (every second) waiting to be able
to obtain an outgoing IP address.  Terminate after an arbitrary amount
of time (90 seconds).