Fixes CC-922: Wait for network for outbound IP
When our appliances (possibly other hosts?) are rebooted, serviced often
starts running before IP addresses have been assigned to all interfaces.
So wait (up to 90 seconds) for one of the interfaces of the type we need
to become available.

Also noticed some confusion between the definition and use of ENDPOINT
and OUTBOUND_IP, so straightened those out.  (For Master, ENDPOINT is
optional and can fall back to OUTBOUND_IP or dynamic.  For Agent,
ENDPOINT is required.  When just parsing options, don't report any
error.  Wait to enforce the rules when we actually go to start the
daemon.)