CC-1387 Adds Docker log options to serviced
Adds two new config values to serviced:
- DOCKER_LOG_DRIVER / --log-driver
- DOCKER_LOG_CONFIG / --log-config

The driver option is used when creating new containers, and is the equivalent of invoking
    docker run ... --log-driver=$DOCKER_LOG_DRIVER

The config option is a comma-separated list of key=value options that are broken into a list, the equivalent of invoking
    docker run ... --log-opt KEY=VALUE --log-opt KEY=VALUE

The driver option is defaulted to `json-file`, so anything logged to stdout or stderr is routed to a logfile.
The config option is defaulted to `max-file=5,max-size=10m`, so logs are truncated at 10 megabytes, and the most recent five logfiles are stored uncompressed on the container host.

These options in place provide a blanket option to prevent services from overrunning storage on /var/lib/docker between logrotate jobs.
These settings apply to all services, and first serviced, then the impacted services, must be restarted for changes to these options to have effect.

Closes CC-1387