Avoid access to uninitialized endpoint variable
Changed endpoints list variable from a slice to a map[int], allowing it
to be sparse in the case that the endpoint is not in the registry.

Previously, the continue statement would cause the endpoints list to
have default data.  This would cause the check for conflicts between
exported and imported endpoint to return a false negative, which
resulted in an erroneous endpoint to be imported.  Serviced would then
open the port (as an imported endpoint), preventing the service
process from doing so.