CC-1896: New private instantiator for RWLock
The coordinator/client/zookeeper package has one case where it works
with objects outside the scope of its Connection, so RWLock needs to
provide a way to lock objects outside the scope of the Connection.

Also found a function that took a RWLock as an input argument (instead
of being a method on RWLock).