ZK leader fix: Avoid spinning on GetW error
Behavior observed: In the TestLeader test, the second entity was supposed
to wait on the lock file of the first entity (the elected leader) then
try to take the lead. Instead it was trying to wait on the parent node
with a trailing /, getting ErrNoNode from GetW(), and going back to
figuring out again who the leader was... So lots of unnecessary
Zookeeper interactions in a tight loop.

Cause: The prevSeq variable was defined as a uint, so it was initialized
to 0. So once the first leader had taken charge, the first entity to
wait in line (seq == 1) would not get prevSeqPath set (because the
leader's 0 was not less than prevSeq == 0). Over time, I believe
entities would eventually wait on the proper lock file (once we got a
pair of s and seq values greater than 0).

Fix: Initialize prevSeq to -1, which meant changing the sequence number
variables to be ints rather than uints. (Which is okay because
Zookeeper's sequence numbers are int32s anyway--see:
http://zookeeper.apache.org/doc/trunk/zookeeperProgrammers.html#Sequence+Nodes+--+Unique+Naming).