CC-2113 CC Master timeout if ZK quorum can't be established
Changed the zookeeper healthcheck to use 'ruok' and 'stat' commands
instead of go-zookeeper connection. This way we can check separately
that the zookeeper agent is up and that quorum is established. Should
zookeeper fail to establish quorum, we can log accordingly.

Iservices that fail their initial healthcheck should now dump the last
10000 lines of their container before shutdown.