Add --register arg to "add host" command
Using setting this boolean command line arg will cause the key to
automatically be installed on the added host.  If the added host is the
local host, the key will simply be written into the proper location; if
it is a remote host a ssh will be opened to run 'serviced host register'
with the key written to stdin.