PoolAPITest maintains internal list of pools
The PoolAPITest mock now maintains an internal list of pools which is
modified by calls to Add and Remove.  Because the API interface does not
use pointer receivers, the list of pools is stored as a pointer to a
slice of pools.  The default list of pools is copied each time in
DefaultPoolAPI().