Allow a non-host master to issue RPC calls.
This lets a master that has not yet been added as a host make
authenticated RPC calls.  This is necessary to allow a master to add
itself as a host.

We achieve this by falling back to a token that contains the master's
public key and is signed by the master's private key.  Such a token can
only be created by the master, since it is the only host that has
access to the master private key.