Fix race condition in authcodec
Allowing the wrapped codec to read and write to the same
data stream we use for authentication results in a race
where the wrapped codec may read the auth header from the
next request while processing the current request.  We fix
this by providing a buffer for the wrapped codec to access
and controlling the data that ends up on that buffer.

An additional test was added that stresses this condition
with multiple simultaneous requests from a single client.