Make path to master key readable
Because we do not have a proper serviced user group, we made the master
host key file world readable.  However, the directory in which that file
exists is not world readable, so the file cannot be read by other
processes (CLI commands in particular).  Make the directory readable.

Per Ian.