Fix race condition in cli/cmd unit test
Previously, the assertion that all expectations were met (i.e., all
routines which were expected to be called were actually called) would
fail for the "WatchConfigFile" routine.  This was due to the fact that
the routine was called in a goroutine, so it may not have been called by
the time it was checked.

This change adds a channel, calls a send on the channel when the mock
routine is called, and waits (with a timeout) on a receive for the
channel.  This ensures that the routine runs before we check the calls.