Refactor "pipe" to take a function with no args
Previously, cmd.pipe() took as arguments a function with a variadic
string parameter, and a variadic string parameter.  The function was
called with the strings.

After this change, the pipe function (which has been renamed to
captureStdout) takes a single parameterless function instead.  This
prevents the proliferation of pipe functions, each taking different
arguments (e.g., pipeAPI).  To pass parameters to the function, a
closure can be formed from an anonymous function.

Also created a captureStderr which operates in a parallel fashion, and
refactored pipeStderr to take a single parameterless function and use
captureStderr.

This design makes it easy to capture the output streams from an
arbitrary function.