Remove race condition from unit tests
Explicitly call routines "at" the desired time in order to test
functionality, rather than starting the interval-based cleanup routine
and using sleep to try to get the proper ordering.  Previously, rare
failures were occurring in these tests.  As a side effect, by removing
the calls to sleep, this change speeds up the test execution by 24
seconds.

Additionally, none of the tests start up the cleanup routine; they do
not need it, and it is unneccesary overhead.