Add Master.UpdateHost to the non-admin RPC list
When delegates check in with the master, they call Master.UpdateHost to
update their properties (e.g., RAM, serviced version, etc.)  Previously,
hosts in a non-admin pool were no able to do so.

Note that this opens a potential security hole which will need to be
addressed in the future.