Fix filesystem-used metric
Previously, the value stored in 'storage.filesystem.used.<TENANT_ID>'
metric was calculated as the difference between the total bytes for
the filesystem and the available bytes; however that is incorrect
because it does not take into account blocks reserved for the
super-user (by default 5%).

After this change the metric will contain the value returned by 'df'.