Remove unecessary ipExists() check when adding a delegate
The ipExists() check is a weak second check to make sure the
delegate has the ip it's being added as.  The authentication
token should verify the delegate now.  This check can be
removed, allowing a user to add a delegate using a NAT ip whose
ports are forwarded to an internal ip address.