Leave DM_BASESIZE unspecified if not given in conf.
This change sets DM_BASESIZE (or "dm.basesize") as an unspecified option to allow code to determine whether a default base size value should be used.  Having DM_BASESIZE automatically acquire a value when it's not given in the config file means that other code has to assume that the value is specified and not a default.

Also changed the associated message when a specified DM_BASESIZE is less than the current base size so that displays values in a multiple of 1024.

Fixes CC-3522.