Don't trim the instance list.
Because we're using the InstanceID as the key in the array, we can
get into a situation where we're splicing out the instance data. We
can see this if we have 2 instances (0 and 1), and we kill instance
0.  We end up with only instance 1 remaining.  When we add the iid
1, the array expands to length = 2, where the instances variable
is instances = [0: undefined, 1: {instance1}].  Our splice then
chops off 1 and we get bad data in the instances table.  We need to
keep all of the instances, and just not show the undefined items
because there's no way to get rid of the 0th item while keeping
the instance while indexing by 1.