Add UI for multiple ZooKeeper instances
A grid was added to the internal ZooKeeper page which contains
information about the different ZooKeeper instances.  Health
checks will show if the ZooKeeper instances are running and if
they are in a quorum.  The grid will also show which instance is
the leader and the number of connections for each instance.