Add the close_timeout setting to filebeat
https://jira.zenoss.com/browse/CC-3893

When the harvester/logstash can't keep up and there's high log
activity resulting in log file rotation, filebeat doesn't release
the file descriptor.  This prevents the disk space from being
recovered and fills the volume.

With the "close_timeout 15m" setting added, the harvester will exit
after 15 minutes.  If there's additional disk activity for the file,
the harvester will resume where it left off.  The only disadvantage
of this is if the harvester stops in the middle of a multiline
message, only the second (resumed) part of the multiline message
will get logged to logstash.  All file descriptors are properly
released.