CC-3490: update klauspost/pgzip and klauspost/compress libs
If during CC backup process we will get into the state where there
will be no space left on the disk the backup process will hang and
never release DFS lock, so new backups won't start. We are checking
space on the disk before doing a backup but if the disk gets full
during the backup process we will just hang.
The problem is in github.com/klauspost/pgzip which we use to zip
files concurrently. The bug is in Writer since it will block when
there is an IOException. Basically goroutines that are performing
zipping will get in a deadlock state.